+++ blosc/__init__.py
@@ -19,6 +19,7 @@ from blosc.toplevel import (
     detect_number_of_cores,
     free_resources,
     set_nthreads,
+    compressor_list,
     print_versions,
     )
 
@@ -48,4 +49,4 @@ from blosc.test import run as test
 __all__ = ['compress', 'compress_ptr', 'decompress', 'decompress_ptr',
            'pack_array', 'unpack_array',
            'detect_number_of_cores', 'free_resources', 'set_nthreads',
-           'print_versions', 'test']
+           'compressor_list', 'print_versions', 'test']